% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenvalues}
\alias{eigenvalues}
\title{Positive eigenvalues of the sample covariance matrix}
\usage{
eigenvalues(x, y)
}
\arguments{
\item{x}{Dataframe or matrix containing the data collected from 1st population.}

\item{y}{Dataframe or matrix containing the data collected from 2nd population (default: \code{NULL}).}
}
\value{
A numeric vector with the positive eigenvalues of the sample covariance matrix or sample pooled covariance matrix when a second dataset is provided.
}
\description{
\code{eigenvalues} computes the positive eigenvalues of the sample covariance matrix using the \code{RcppArmadillo} interface for fast eigen-decomposition via divide and conquer method.
}
\keyword{internal}

