% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_power}
\alias{matrix_power}
\title{Power of symmetric positive definite matrix}
\usage{
matrix_power(matrix, power)
}
\arguments{
\item{matrix}{Symmetric positive definite matrix to compute the power from.}

\item{power}{Desired power.}
}
\value{
A symmetric positive definite matrix with same eigenvectors as input
  matrix and eigenvalues being the powered eigenvalues of the input matrix.
}
\description{
\code{matrix_power} computes the power of a symmetric positive definite
matrix using spectral decomposition via the \code{eigen} function of the R
\code{base} package in its symmetric version.
}
\keyword{internal}
