% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_testing.R
\name{parametric_test}
\alias{parametric_test}
\title{Hotelling's parametric test for the mean}
\usage{
parametric_test(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0,
  skip_check = FALSE)
}
\arguments{
\item{x}{Dataframe or matrix containing the data collected from 1st
population.}

\item{y}{Dataframe or matrix containing the data collected from 2nd
population (default: \code{NULL}).}

\item{mu}{True mean value (or difference between mean values) under the null
hypothesis (default = \code{0}).}

\item{paired}{Is the input data paired? (default: \code{FALSE}).}

\item{step_size}{The step size used to perform integral approximation via the
method of rectangles (default: \code{0}). When set to \code{0}, it assumes
that we are dealing with multivariate data rather than functional data and
thus no integration is necessary.}

\item{skip_check}{Flag to skip checks on input types. Default is off. This
parameter should always be set to its default \code{FALSE}. It is set to
\code{TRUE} only internally, in power calculation functions, in which data
generated by the hypothesized model is already in matrix form.}
}
\value{
A 3-column \code{\link[dplyr]{data_frame}} containing the names of
  the required statistics and their values, along with an estimate of the
  p-value of the corresponding statistical test. This function uses
  exclusively Hotelling's T^2 statistic. Hence, the output
  \code{\link[dplyr]{data_frame}} only has 1 row displaying information
  about Hotelling's statistic.
}
\description{
\code{parametric_test} performs a parametric test for the mean (or for the
difference between the means) of populations of curves based on Hotelling's
T^2 statistic.
}
\examples{
x <- aneurisk \%>\%
 dplyr::filter(variable == "curvature" & group == "low") \%>\%
 dplyr::select(data) \%>\%
 purrr::flatten_df()
y <- aneurisk \%>\%
 dplyr::filter(variable == "curvature" & group == "up") \%>\%
 dplyr::select(data) \%>\%
 purrr::flatten_df()
step_size <- aneurisk$abscissa[[1]][2] - aneurisk$abscissa[[1]][1]
parametric_test(x = x, y = y, mu = 0, step_size = step_size)
}
\seealso{
This test is described in details in Secchi, P., Stamm, A., &
  Vantini, S. (2013). \emph{Inference for the mean of large \eqn{p} small
  \eqn{n} data: A finite-sample high-dimensional generalization of
  Hotelling theorem}. Electronic Journal of Statistics, 7, pp. 2005-2031.
  doi:10.1214/13-EJS833, available online at
  \url{http://projecteuclid.org/euclid.ejs/1375708877}.
}
