% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition}
\alias{partition}
\title{Data partitioning}
\usage{
partition(df, n, probs)
}
\arguments{
\item{df}{Input dataset to be partitioned.}

\item{n}{Number of random partitions.}

\item{probs}{A named vector of probabilities defining the names and
relative sizes of each subsequent subgroup.}
}
\value{
A character vector of size \code{n} in which each entry contains the
name of one of the subgroups defined in the \code{probs} vector.
}
\description{
\code{partition} generates random partitions of the input dataset into
smaller groups.
}
\keyword{internal}
