% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_testing.R
\name{permutation_test}
\alias{permutation_test}
\title{Permutation test for the mean}
\usage{
permutation_test(x, y = NULL, mu = 0, paired = FALSE, step_size = 0,
  B = 1000, statistic = "Hotelling", verbose = TRUE, skip_check = FALSE)
}
\arguments{
\item{x}{Dataframe or matrix containing the data collected from 1st
population.}

\item{y}{Dataframe or matrix containing the data collected from 2nd
population (default: \code{NULL}).}

\item{mu}{True mean value (or difference between mean values) under the null
hypothesis (default = \code{0}).}

\item{paired}{Is the input data paired? (default: \code{FALSE}).}

\item{step_size}{The step size used to perform integral approximation via the
method of rectangles (default: \code{0}). When set to \code{0}, it assumes
that we are dealing with multivariate data rather than functional data and
thus no integration is necessary.}

\item{B}{Number of bootstrap permutations (default: 1000).}

\item{statistic}{Statistic to be used within the permutation framework.
Choices are Hotelling (default), L1, L2, Linf, StandardizedL1,
StandardizedL2, StandardizedLinf and All.}

\item{verbose}{Activate verbose (default: \code{TRUE}).}

\item{skip_check}{Flag to skip checks on input types. Default is off. This
parameter should always be set to its default \code{FALSE}. It is set to
\code{TRUE} only internally, in power calculation functions, in which data
generated by the hypothesized model is already in matrix form.}
}
\value{
A 3-column \code{\link[dplyr]{data_frame}} containing the names of
  the required statistics and their values, along with an estimate of the
  p-value of the corresponding statistical test.
}
\description{
\code{permutation_test} performs a non-parametric test for the mean function
 of a single population of curves or for the difference between mean
 functions of two populations of curves. The one-population test is based on
 central symmetry and reflexions around the mean function while the
 two-population test rely on permutations.
}
\details{
All the statistics of the \pkg{fdahotelling} package can be embedded in this
 test. In addition, the user can implement its own statistic to be used with
 the test, provided that:
 \itemize{
 \item the user-defined statistic takes the same input arguments as the ones
   in the \pkg{fdahotelling} package, i.e., the mandatory first dataset
   \code{x}, the optional second dataset \code{y} and the mean function under
   the null hypothesis \code{mu} that can have a default value,
 \item high values of the statistic yield rejection of the null hypothesis.
 }
}
\examples{
x <- aneurisk \%>\%
 dplyr::filter(variable == "curvature" & group == "low") \%>\%
 dplyr::select(data) \%>\%
 purrr::flatten_df()
y <- aneurisk \%>\%
 dplyr::filter(variable == "curvature" & group == "up") \%>\%
 dplyr::select(data) \%>\%
 purrr::flatten_df()
step_size <- aneurisk$abscissa[[1]][2] - aneurisk$abscissa[[1]][1]
permutation_test(x = x, y = y, mu = 0,
 statistic = "All", step_size = step_size)
}
\seealso{
This test is described in details in the technical report by Pini,
  A., Stamm, A., & Vantini, S. (2015). \emph{Hotelling \eqn{T^2} in
  functional Hilbert spaces}, available online at
  \url{https://mox.polimi.it/publication-results/?id=524&tipo=add_qmox}.
}
