% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pseudoinverse}
\alias{pseudoinverse}
\title{Pseudoinverse of the sample covariance matrix}
\usage{
pseudoinverse(x, y, tolerance = 0)
}
\arguments{
\item{x}{Dataframe or matrix containing the data collected from 1st population.}

\item{y}{Dataframe or matrix containing the data collected from 2nd population (default: \code{NULL}).}

\item{tolerance}{Positive value \eqn{\tau} such that positive eigenvalues of the sample covariance matrix that are below \eqn{\tau} tr\eqn{\Sigma} are discarded for inverse computation (default: 0).}
}
\value{
A matrix containing the pseudoinverse of the sample covariance matrix.
}
\description{
\code{pseudoinverse} computes the Moore-Penrose inverse of the sample covariance matrix using the \code{RcppArmadillo} interface for fast pseudoinverse calculations via divide and conquer method.
}
\keyword{internal}
