% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_group}
\alias{random_group}
\title{Random index generator for data splitting}
\usage{
random_group(n, probs)
}
\arguments{
\item{n}{Total number of indices to be splitted into smaller groups.}

\item{probs}{A named vector of probabilities defining the names and
proportions of indices defining each subgroup.}
}
\value{
A character vector of size \code{n} in which each entry contains the
name of one of the subgroups defined in the \code{probs} vector.
}
\description{
\code{random_group} randomly splits a vector of indices (defined by the
number of lines of a dataset) into subgroups.
}
\keyword{internal}

