% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rmvnorm}
\alias{rmvnorm}
\title{Multivariate Gaussian random generator in C++}
\usage{
rmvnorm(n, mean, squareRootSigma)
}
\arguments{
\item{n}{Number of samples to be drawn from the distribution (default: 1).}

\item{mean}{Mean vector of the distribution (default: 0).}

\item{squareRootSigma}{Square-root of the variance-covariance matrix of the distribution (default = 1).}
}
\value{
A matrix of size \eqn{n\times} \code{length(mean)} containing \eqn{n} randomly sampled vectors of size \code{length(mean)}.
}
\description{
\code{rmvnorm} generates random samples from a multivariate Gaussian distribution with mean \eqn{\mu} and variance-covariance matrix \eqn{\Sigma} assumed positive semi-definite.
}
\keyword{internal}

