% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stat_hotelling_impl}
\alias{stat_hotelling_impl}
\title{Compute Hotelling's \eqn{T^2} statistic}
\usage{
stat_hotelling_impl(x, y, mu, paired = FALSE, step_size = 0,
  use_correction = FALSE, tolerance = 0)
}
\arguments{
\item{x}{Dataframe or matrix containing the data collected from 1st
population.}

\item{y}{Dataframe or matrix containing the data collected from 2nd
population (default: \code{NULL}).}

\item{mu}{True mean value (or difference between mean values) under the null
hypothesis (default = \code{0}).}

\item{paired}{Is the input data paired? (default: \code{FALSE}).}

\item{step_size}{The step size used to perform integral approximation via the
method of rectangles (default: \code{0}). When set to \code{0}, it assumes
that we are dealing with multivariate data rather than functional data and
thus no integration is necessary.}

\item{use_correction}{Flag to turn on and off a statistic adjustment factor that makes it Fisher-distributed under the normality assumption (default: \code{FALSE}).}

\item{tolerance}{Positive value \eqn{\tau} such that positive eigenvalues of the sample covariance matrix that are below \eqn{\tau} tr\eqn{\Sigma} are discarded for inverse computation (default: 0).}
}
\value{
A named numeric vector of size 1 storing the Hotelling's \eqn{T^2} statistic defined as \deqn{\int_D}.
}
\description{
\code{stat_hotelling_impl} computes the generalized Hotelling's \eqn{T^2} statistic using the \code{RcppArmadillo} interface for fast pseudoinverse calculations via divide and conquer method.
}
\keyword{internal}
