% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\alias{stat_hotelling}
\alias{stat_L1}
\alias{stat_L1_std}
\alias{stat_L2}
\alias{stat_L2_std}
\alias{stat_Linf}
\alias{stat_Linf_std}
\alias{stat_all}
\title{Statistics for functional data}
\usage{
stat_hotelling(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)

stat_L1(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)

stat_L1_std(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)

stat_L2(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)

stat_L2_std(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)

stat_Linf(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)

stat_Linf_std(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)

stat_all(x, y = NULL, mu = NULL, paired = FALSE, step_size = 0)
}
\arguments{
\item{x}{Dataframe or matrix containing the data collected from 1st
population.}

\item{y}{Dataframe or matrix containing the data collected from 2nd
population (default: \code{NULL}).}

\item{mu}{True mean value (or difference between mean values) under the null
hypothesis (default = \code{0}).}

\item{paired}{Is the input data paired? (default: \code{FALSE}).}

\item{step_size}{The step size used to perform integral approximation via the
method of rectangles (default: \code{0}). When set to \code{0}, it assumes
that we are dealing with multivariate data rather than functional data and
thus no integration is necessary.}
}
\value{
All \code{stat_*} functions return a named numeric vector of size 1
  storing the value of the corresponding statistic except \code{stat_all}
  that outputs a vector of size 7 containing the values of all 7 statistics
  currently available in the \pkg{fdahotelling} package.
}
\description{
Compute various statistics for functional data using numerical integration
when needed. A total of 7 statistics are available in the package and the
user can even define its own statistic for use within the statistical test
functions.
}
\seealso{
A detailed description of the individual statistics for functional
  data provided by the \pkg{fdahotelling} package can be found in the vignette
  \emph{Available statistics for functional data}.
}
