% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{tests}
\alias{tests}
\alias{test_onesample}
\alias{test_twosample}
\title{Inference on the Mean for Multivariate and Functional Data}
\usage{
test_onesample(x, mu = 0, step_size = 0, B = 1000,
  statistic = "Hotelling", verbose = TRUE, skip_check = FALSE)

test_twosample(x, y, mu = 0, paired = FALSE, step_size = 0, B = 1000,
  statistic = "Hotelling", verbose = TRUE, skip_check = FALSE)
}
\arguments{
\item{x}{Dataframe or matrix containing the data collected from 1st
population.}

\item{mu}{True mean value (or difference between mean values) under the null
hypothesis (default = \code{0}).}

\item{step_size}{The step size used to perform integral approximation via the
method of rectangles (default: \code{0}). When set to \code{0}, it assumes
that we are dealing with multivariate data rather than functional data and
thus no integration is necessary.}

\item{B}{Number of bootstrap permutations (default: 1000).}

\item{statistic}{Statistic to be used within the permutation framework.
Choices are Hotelling (default), L1, L2, Linf, StandardizedL1,
StandardizedL2, StandardizedLinf and All.}

\item{verbose}{Activate verbose (default: \code{TRUE}).}

\item{skip_check}{Flag to skip checks on input types. Default is off. This
parameter should always be set to its default \code{FALSE}. It is set to
\code{TRUE} only internally, in power calculation functions, in which data
generated by the hypothesized model is already in matrix form.}

\item{y}{Dataframe or matrix containing the data collected from 2nd
population (default: \code{NULL}).}

\item{paired}{Is the input data paired? (default: \code{FALSE}).}
}
\value{
A 3-column \code{\link[tibble]{tibble}} containing the names of the
  required statistics and their values, along with an estimate of the p-value
  of the corresponding statistical test.
}
\description{
These functions implement testing procedures for the mean of populations of
vectors or functions. In particular, when the optional parameter \code{B =
0}, a parametric test based on Hotelling's statistic is used, while, if
\code{B > 0}, a permutation test is instead used and more statistics can be
chosen by the user. The parametric test assumes that the data is normally
distributed or that the sample size is large such that the central limit
theorem can be applied. The permutation test assumes that the distribution
from which the data come is symmetric.
}
\details{
All the statistics of the \pkg{fdahotelling} package can be used with the
permutation version of this test. In addition, the user can implement its own
statistic to be used with the test, provided that: \itemize{ \item the
user-defined statistic takes the same input arguments as the ones in the
\pkg{fdahotelling} package, i.e., the mandatory first dataset \code{x}, the
optional second dataset \code{y} and the mean function under the null
hypothesis \code{mu} that can have a default value, \item high values of the
statistic yield rejection of the null hypothesis. }
}
\examples{
####################
# One-Sample Tests #
####################
x <- aneurisk \%>\%
 dplyr::filter(variable == "curvature" & group == "low") \%>\%
 dplyr::select(data) \%>\%
 purrr::flatten_df()
# Parametric test
test_onesample(x = x, mu = 0, step_size = 0.01, B = 0)
# Permutation test
test_onesample(x = x, mu = 0, step_size = 0.01, B = 100)
####################
# Two-Sample Tests #
####################
y <- aneurisk \%>\%
 dplyr::filter(variable == "curvature" & group == "up") \%>\%
 dplyr::select(data) \%>\%
 purrr::flatten_df()
# Parametric test
test_twosample(x = x, y = y, mu = 0, step_size = 0.01, B = 0)
# Permutation test
test_twosample(x = x, y = y, mu = 0, step_size = 0.01, B = 100)

}
\seealso{
The parametric test is described in details in Secchi, P., Stamm,
  A., & Vantini, S. (2013). \emph{Inference for the mean of large \eqn{p}
  small \eqn{n} data: A finite-sample high-dimensional generalization of
  Hotelling theorem}. Electronic Journal of Statistics, 7, pp. 2005-2031.
  doi:10.1214/13-EJS833, available online at
  \url{http://projecteuclid.org/euclid.ejs/1375708877}. The permutation test
  is described in details in the technical report by Pini, A., Stamm, A., &
  Vantini, S. (2015). \emph{Hotelling \eqn{T^2} in functional Hilbert
  spaces}, available online at
  \url{https://mox.polimi.it/publication-results/?id=524&tipo=add_qmox}.
}
